[server] Introduce EntitlementServiceChargbee and move parts of EligibilityService into it
